@@ -1,64 +1,69 @@
 {
-  "name": "aurora-launcher",
-  "version": "0.0.2",
-  "description": "Launcher for Minecraft",
-  "main": "build/main/index.js",
-  "scripts": {
-    "dev": "npm run build && npm run start:dev",
-    "build": "webpack",
-    "start:dev": "cross-env DEV=true electron .",
-    "start:prod": "electron .",
-    "clear": "rimraf build aurora-launcher-win32-x64",
-    "pack": "electron-packager build"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/AuroraTeam/Launcher.git"
-  },
-  "author": "AuroraTeam",
-  "contributors": [
-    "JoCat (https://github.com/JoCat)"
-  ],
-  "license": "GPL-3.0",
-  "bugs": {
-    "url": "https://github.com/AuroraTeam/Launcher/issues"
-  },
-  "homepage": "https://github.com/AuroraTeam/Launcher#readme",
-  "devDependencies": {
-    "@sweetalert2/theme-dark": "^4.0.3",
-    "@types/semver": "^7.3.4",
-    "@types/ws": "^7.4.0",
-    "aurora-api": "^0.1.3",
-    "copy-webpack-plugin": "^8.1.0",
-    "cross-env": "^7.0.3",
-    "css-loader": "^5.2.0",
-    "electron": "^12.0.2",
-    "electron-devtools-installer": "git+https://github.com/MarshallOfSound/electron-devtools-installer.git#9fad2da5a6961c80f83fc9d1ca601af59a1ae37a",
-    "electron-packager": "^15.2.0",
-    "file-loader": "^6.2.0",
-    "html-minimizer-webpack-plugin": "^2.1.0",
-    "less": "^4.1.1",
-    "less-loader": "^8.0.0",
-    "rimraf": "^3.0.2",
-    "sass": "^1.32.8",
-    "sass-loader": "^11.0.1",
-    "style-loader": "^2.0.0",
-    "stylus": "^0.54.8",
-    "stylus-loader": "^5.0.0",
-    "sweetalert2": "^10.15.7",
-    "ts-loader": "^8.0.18",
-    "typescript": "^4.2.3",
-    "vue": "^2.6.12",
-    "vue-loader": "^15.9.6",
-    "vue-router": "^3.5.1",
-    "vue-sweetalert2": "^4.2.1",
-    "vue-template-compiler": "^2.6.12",
-    "webpack": "^5.28.0",
-    "webpack-cli": "^4.5.0"
-  },
-  "dependencies": {
-    "p-map": "^4.0.0",
-    "semver": "^7.3.5",
-    "source-map-support": "^0.5.19"
-  }
+    "name": "aurora-launcher",
+    "version": "0.0.2",
+    "description": "Launcher for Minecraft",
+    "main": "build/main/index.js",
+    "scripts": {
+        "dev": "npm run build:main && concurrently \"webpack serve\" \"npm:start:dev\" ",
+        "build": "npm run build:main && npm run build:renderer",
+        "build:main": "esbuild src/main/index.ts --platform=node --bundle --outdir=build/main --sourcemap=inline --external:electron",
+        "build:renderer": "webpack",
+        "start:dev": "cross-env DEV=true electron .",
+        "start:prod": "electron .",
+        "clear": "rimraf build aurora-launcher-win32-x64",
+        "pack": "electron-packager build"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/AuroraTeam/Launcher.git"
+    },
+    "author": "AuroraTeam",
+    "contributors": [
+        "JoCat (https://github.com/JoCat)"
+    ],
+    "license": "GPL-3.0",
+    "bugs": {
+        "url": "https://github.com/AuroraTeam/Launcher/issues"
+    },
+    "homepage": "https://github.com/AuroraTeam/Launcher#readme",
+    "devDependencies": {
+        "@sweetalert2/theme-dark": "^5.0.2",
+        "@types/semver": "^7.3.8",
+        "@types/ws": "^7.4.7",
+        "aurora-api": "^0.1.3",
+        "concurrently": "^6.2.1",
+        "copy-webpack-plugin": "^9.0.1",
+        "cross-env": "^7.0.3",
+        "css-loader": "^6.2.0",
+        "electron": "^13.2.2",
+        "electron-devtools-installer": "^3.2.0",
+        "electron-packager": "^15.3.0",
+        "file-loader": "^6.2.0",
+        "html-minimizer-webpack-plugin": "^3.1.1",
+        "less": "^4.1.1",
+        "less-loader": "^10.0.1",
+        "rimraf": "^3.0.2",
+        "sass": "^1.38.1",
+        "sass-loader": "^12.1.0",
+        "style-loader": "^3.2.1",
+        "stylus": "^0.54.8",
+        "stylus-loader": "^6.1.0",
+        "sweetalert2": "^11.1.4",
+        "ts-loader": "^9.2.5",
+        "typescript": "^4.3.5",
+        "vue": "^2.6.14",
+        "vue-loader": "^15.9.8",
+        "vue-router": "^3.5.2",
+        "vue-sweetalert2": "^5.0.2",
+        "vue-template-compiler": "^2.6.14",
+        "webpack": "^5.51.1",
+        "webpack-cli": "^4.8.0"
+    },
+    "dependencies": {
+        "esbuild": "^0.12.24",
+        "p-map": "^5.1.0",
+        "semver": "^7.3.5",
+        "source-map-support": "^0.5.19",
+        "webpack-dev-server": "^4.0.0"
+    }
 }
