@@ -1,25 +1,27 @@
-import { AuroraAPI } from 'aurora-api'
-import { App } from '..'
-const apiConfig = require('@config').api
+import { AuroraAPI } from 'aurora-api';
+import { App } from '..';
+const apiConfig = require('../../../config.json').api;
 
 export default class APIManager {
-    api: AuroraAPI | null = null
+    api: AuroraAPI | null = null;
 
     constructor() {
-        this.connect()
+        this.connect();
     }
 
     async connect() {
         try {
-            this.api = await new AuroraAPI().connect(apiConfig.url || 'ws://localhost:1370') as AuroraAPI
-            App.window.sendEvent('apiConnectSuccess')
+            this.api = (await new AuroraAPI().connect(
+                apiConfig.url || 'ws://localhost:1370'
+            )) as AuroraAPI;
+            App.window.sendEvent('apiConnectSuccess');
         } catch (error) {
-            App.window.sendEvent('apiConnectError', 'Ошибка при подключении')
-            console.log(error)
+            App.window.sendEvent('apiConnectError', 'Ошибка при подключении');
+            console.log(error);
         }
     }
 
     public async send(type: string, data?: object) {
-        return await this.api?.send(type, data)
+        return await this.api?.send(type, data);
     }
 }
